Reimplement LoadRepository as RepositoryAtCommit using pygit2
